requirements.txt
# =========================
# Core Backend Framework
# =========================
fastapi==0.110.0
uvicorn[standard]==0.27.1

# =========================
# Data Handling & Analysis
# =========================
pandas==2.2.2
numpy==1.26.4

# =========================
# Garmin Data Parsing
# =========================
fitparse==1.2.0

# =========================
# Validation & Schemas
# =========================
pydantic==2.6.4
python-multipart==0.0.9

# =========================
# Environment Variables
# =========================
python-dotenv==1.0.1

# =========================
# Scientific / Signal Processing
# =========================
scipy==1.12.0

# =========================
# AI / Chat Layer (Phase-1 ready)
# =========================
openai==1.30.1
tiktoken==0.7.0

# =========================
# Logging & Utilities
# =========================
loguru==0.7.2

# =========================
# Testing
# =========================
pytest==8.1.1
httpx==0.27.0

# =========================
# Optional (Future / Phase-2)
# =========================
scikit-learn==1.4.2
joblib==1.4.2

# =========================
# Deployment (Optional)
# =========================
gunicorn==21.2.0
